pyproject.toml
[project]
name = "advanced-ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
